Expose connection's streaminfo to the HTTP Lua filter
This allows Lua to access the connection StreamInfo, and thus the
dynamic metadata. This makes it possible for HTTP requests to act on
the Cloud provider specific PROXY PROTOCOL TLV extensions.

The bulk of the lines of this are moving the StreamInfoWrapper from
the Lua HTTP directory into the common Lua directory and fixing up the
references.

This uses an (unfortunate) const_cast to match the existing semantics
of StreamInfoWrapper.  The alternatives are much more invasive at the
moment.
