+++ source/extensions/filters/common/lua/wrappers.cc
@@ -379,6 +379,137 @@ int ConnectionWrapper::luaSsl(lua_State* state) {
   return 1;
 }
 
+int ConnectionWrapper::luaStreamInfo(lua_State* state) {
+  auto& streaminfo = connection_->streamInfo();
+  if (stream_info_wrapper_.get() != nullptr) {
+    stream_info_wrapper_.pushStack();
+  } else {
+    stream_info_wrapper_.reset(Filters::Common::Lua::StreamInfoWrapper::create(state, streaminfo),
+                               true);
+  }
+  return 1;
+}
+
+int StreamInfoWrapper::luaProtocol(lua_State* state) {
+  const std::string& protocol =
+      Envoy::Http::Utility::getProtocolString(stream_info_.protocol().value());
+  lua_pushlstring(state, protocol.data(), protocol.size());
+  return 1;
+}
+
+int StreamInfoWrapper::luaDynamicMetadata(lua_State* state) {
+  if (dynamic_metadata_wrapper_.get() != nullptr) {
+    dynamic_metadata_wrapper_.pushStack();
+  } else {
+    dynamic_metadata_wrapper_.reset(DynamicMetadataMapWrapper::create(state, *this), true);
+  }
+  return 1;
+}
+
+int StreamInfoWrapper::luaDownstreamSslConnection(lua_State* state) {
+  const auto& ssl = stream_info_.downstreamAddressProvider().sslConnection();
+  if (ssl != nullptr) {
+    if (downstream_ssl_connection_.get() != nullptr) {
+      downstream_ssl_connection_.pushStack();
+    } else {
+      downstream_ssl_connection_.reset(
+          Filters::Common::Lua::SslConnectionWrapper::create(state, *ssl), true);
+    }
+  } else {
+    lua_pushnil(state);
+  }
+  return 1;
+}
+
+int StreamInfoWrapper::luaDownstreamLocalAddress(lua_State* state) {
+  const std::string& local_address =
+      stream_info_.downstreamAddressProvider().localAddress()->asString();
+  lua_pushlstring(state, local_address.data(), local_address.size());
+  return 1;
+}
+
+int StreamInfoWrapper::luaDownstreamDirectRemoteAddress(lua_State* state) {
+  const std::string& direct_remote_address =
+      stream_info_.downstreamAddressProvider().directRemoteAddress()->asString();
+  lua_pushlstring(state, direct_remote_address.data(), direct_remote_address.size());
+  return 1;
+}
+
+int StreamInfoWrapper::luaDownstreamRemoteAddress(lua_State* state) {
+  const std::string& remote_address =
+      stream_info_.downstreamAddressProvider().remoteAddress()->asString();
+  lua_pushlstring(state, remote_address.data(), remote_address.size());
+  return 1;
+}
+
+int StreamInfoWrapper::luaRequestedServerName(lua_State* state) {
+  absl::string_view requested_serve_name =
+      stream_info_.downstreamAddressProvider().requestedServerName();
+  lua_pushlstring(state, requested_serve_name.data(), requested_serve_name.size());
+  return 1;
+}
+
+DynamicMetadataMapIterator::DynamicMetadataMapIterator(DynamicMetadataMapWrapper& parent)
+    : parent_{parent}, current_{parent_.streamInfo().dynamicMetadata().filter_metadata().begin()} {}
+
+StreamInfo::StreamInfo& DynamicMetadataMapWrapper::streamInfo() { return parent_.stream_info_; }
+
+int DynamicMetadataMapIterator::luaPairsIterator(lua_State* state) {
+  if (current_ == parent_.streamInfo().dynamicMetadata().filter_metadata().end()) {
+    parent_.iterator_.reset();
+    return 0;
+  }
+
+  lua_pushlstring(state, current_->first.data(), current_->first.size());
+  Filters::Common::Lua::MetadataMapHelper::createTable(state, current_->second.fields());
+
+  current_++;
+  return 2;
+}
+
+int DynamicMetadataMapWrapper::luaGet(lua_State* state) {
+  const char* filter_name = luaL_checkstring(state, 2);
+  const auto& metadata = streamInfo().dynamicMetadata().filter_metadata();
+  const auto filter_it = metadata.find(filter_name);
+  if (filter_it == metadata.end()) {
+    return 0;
+  }
+
+  Filters::Common::Lua::MetadataMapHelper::createTable(state, filter_it->second.fields());
+  return 1;
+}
+
+int DynamicMetadataMapWrapper::luaSet(lua_State* state) {
+  if (iterator_.get() != nullptr) {
+    luaL_error(state, "dynamic metadata map cannot be modified while iterating");
+  }
+
+  const char* filter_name = luaL_checkstring(state, 2);
+  const char* key = luaL_checkstring(state, 3);
+
+  // MetadataMapHelper::loadValue will convert the value on top of the Lua stack,
+  // so push a copy of the 3rd arg ("value") to the top.
+  lua_pushvalue(state, 4);
+
+  ProtobufWkt::Struct value;
+  (*value.mutable_fields())[key] = Filters::Common::Lua::MetadataMapHelper::loadValue(state);
+  streamInfo().setDynamicMetadata(filter_name, value);
+
+  // Pop the copy of the metadata value from the stack.
+  lua_pop(state, 1);
+  return 0;
+}
+
+int DynamicMetadataMapWrapper::luaPairs(lua_State* state) {
+  if (iterator_.get() != nullptr) {
+    luaL_error(state, "cannot create a second iterator before completing the first");
+  }
+
+  iterator_.reset(DynamicMetadataMapIterator::create(state, *this), true);
+  lua_pushcclosure(state, DynamicMetadataMapIterator::static_luaPairsIterator, 1);
+  return 1;
+}
+
 } // namespace Lua
 } // namespace Common
 } // namespace Filters
