+++ source/extensions/filters/http/lua/BUILD
@@ -39,7 +39,6 @@ envoy_cc_library(
     hdrs = ["wrappers.h"],
     deps = [
         "//envoy/http:header_map_interface",
-        "//envoy/stream_info:stream_info_interface",
         "//source/common/crypto:utility_lib",
         "//source/common/http:header_utility_lib",
         "//source/common/http:utility_lib",
