+++ test/extensions/filters/http/ext_authz/config_test.cc
@@ -197,7 +197,7 @@ private:
     Http::TestRequestHeaderMapImpl request_headers;
     NiceMock<Http::MockStreamDecoderFilterCallbacks> decoder_callbacks;
     ON_CALL(decoder_callbacks, connection())
-        .WillByDefault(Return(OptRef<const Network::Connection>{connection_}));
+        .WillByDefault(Return(OptRef<Network::Connection>{connection_}));
     filter->setDecoderFilterCallbacks(decoder_callbacks);
     EXPECT_EQ(Http::FilterHeadersStatus::StopAllIterationAndWatermark,
               filter->decodeHeaders(request_headers, false));
