+++ test/extensions/filters/http/lua/lua_filter_test.cc
@@ -111,7 +111,7 @@ public:
   void setupSecureConnection(const bool secure) {
     ssl_ = std::make_shared<NiceMock<Envoy::Ssl::MockConnectionInfo>>();
     EXPECT_CALL(decoder_callbacks_, connection())
-        .WillOnce(Return(OptRef<const Network::Connection>{connection_}));
+        .WillOnce(Return(OptRef<Network::Connection>{connection_}));
     EXPECT_CALL(Const(connection_), ssl()).WillOnce(Return(secure ? ssl_ : nullptr));
   }
