+++ test/extensions/filters/http/wasm/wasm_filter_test.cc
@@ -1809,7 +1809,7 @@ TEST_P(WasmHttpFilterTest, Property) {
   NiceMock<Network::MockConnection> connection;
   EXPECT_CALL(connection, id()).WillRepeatedly(Return(4));
   EXPECT_CALL(encoder_callbacks_, connection())
-      .WillRepeatedly(Return(OptRef<const Network::Connection>{connection}));
+      .WillRepeatedly(Return(OptRef<Network::Connection>{connection}));
   std::shared_ptr<Router::MockRoute> route{new NiceMock<Router::MockRoute>()};
   EXPECT_CALL(request_stream_info_, route()).WillRepeatedly(Return(route));
   std::shared_ptr<NiceMock<Envoy::Upstream::MockHostDescription>> host_description(
